RES: prefer trait methods before private inherent methods
