+++ src/test/kotlin/org/rust/lang/core/resolve/RsMultiResolveTest.kt
@@ -110,30 +110,6 @@ class RsMultiResolveTest : RsResolveTestBase() {
         }   //^
     """)
 
-    fun `test trait method and private inherent method`() = doTest("""
-        use foo::{Foo, Trait};
-
-        mod foo {
-            pub struct Foo;
-            impl Foo {
-                fn get(&self) { println!("struct"); }
-            }
-
-            pub trait Trait {
-                fn get(&self);
-            }
-            impl Trait for Foo {
-                fn get(&self) { println!("trait"); }
-            }
-        }
-
-        fn main() {
-            let f = foo::Foo;
-            f.get();
-            //^
-        }
-    """)
-
     private fun doTest(@Language("Rust") code: String) {
         InlineFile(code)
         val element = findElementInEditor<RsReferenceElement>()
